Moved lightworks qubit into a dedicated module - This enables "from lightworks import qubit" syntax and provides better separation of these components
