+++ lightworks/__init__.py
@@ -49,11 +49,10 @@ from .sdk.state import State
 from .sdk.utils import random_unitary, random_permutation
 from .sdk.utils import db_loss_to_transmission, transmission_to_db_loss
 from .sdk.optimisation import Optimisation
-from .sdk import qubit
 
 from .sdk.utils.exceptions import *
 
 __all__ = ["Circuit", "Unitary", "Display", "State", "random_unitary", 
            "random_permutation", "db_loss_to_transmission", 
            "transmission_to_db_loss", "Parameter", "ParameterDict", 
-           "Optimisation", "qubit"]
+           "Optimisation", "emulator", "qubit"]
